Mini-Project
I have created a code that will make a CSV file for the School administration and it will take prior input data from the user and then ask for the approval whether the input data is correct or not and then it will either ask to re-enter the data or it will go the next step. Once the data is correct it will ask whether you have to give more inputs or not and further actions will be taken accordingly.
Once you have ended giving all those inputs, a CSV file will be created and then if you open the file in MS Excel , you will find the table being sorted accordingly as you have entered during the input period.